Pimped NewtonFractal
Instantly writes to .gif instead of keeping images in memory while
rendering the entire animation.

When selecting a non-square image size, the image no longer looks
distorted (this was because it used a 'scale' effect, when instead I
wanted a 'crop' effect).

Created labels on images (in the lower left corner) to give some
information about the fractal. (this is because I auto-generate the
images, and want to know which settings make the prettiest fractals).

Fixed some overloading errors with HSV and HSL color spaces.

Fixed some conversion to 24-bit RGB integer.